move Elliptic curve utils into a separate file and add `ecVerify` method with test
